+++ src/HybridClusterWorker.cpp
@@ -12,11 +12,11 @@ HybridClusterWorker::HybridClusterWorker(
   const std::vector<std::string> &seq,
   ClusterAlgorithm &clust_algo,
   const std::uint8_t threads,
-  const double breakpoint,
-  bool verbose
-) : AlignClusterWorker(seq, clust_algo, threads, verbose), breakpoint(breakpoint) {};
+  const double breakpoint
+) : AlignClusterWorker(seq, clust_algo, threads), breakpoint(breakpoint) {};
 
-void HybridSplitClusterWorker::operator()(std::size_t begin, std::size_t end) {
+template<int verbose>
+void HybridSplitClusterWorker<verbose>::operator()(std::size_t begin, std::size_t end) {
   double n = seq.size();
   double m = (n*n - 3.0*n + 2.0)/2.0;
   size_t my_prealigned = 0;
@@ -33,33 +33,32 @@ void HybridSplitClusterWorker::operator()(std::size_t begin, std::size_t end) {
     begin_i = round(1.5 + 0.5*sqrt(9.0 + 8.0*((m*begin)/threads - 1.0)));
   }
   size_t end_i   = round(1.5 + 0.5*sqrt(9.0 + 8.0*((m*end)/threads - 1.0)));
-  if (verbose) {
-    mutex.lock();
-    OPTIMOTU_COUT << "HybridSplit thread " << begin << " entered; sequences [" <<
-      begin_i << ", "<< end_i << ")" << std::endl;
-    mutex.unlock();
-  }
+  OPTIMOTU_DEBUG(
+    1,
+    << "HybridSplit thread " << begin
+    << " entered; sequences [" << begin_i
+    << ", "<< end_i << ")" << std::endl
+  );
   for (size_t i = begin_i; i < end_i; i++) {
     for (size_t j = 0; j < i; j++) {
       double threshold = my_algo->max_relevant(i, j);
-      if (verbose) {
-        mutex.lock();
-        OPTIMOTU_COUT << "seqs " << j
-                      << " and " << i
-                      << " max relevant=" << threshold
-                      << std::endl;
-        mutex.unlock();
-      }
+      OPTIMOTU_DEBUG(
+        2,
+        << "thread" << begin
+        << ": seqs " << j
+        << " and " << i
+        << " max relevant=" << threshold
+        << std::endl
+      );
       bool is_seqj_longer = seq[j].size() > seq[i].size();
       size_t s1 = is_seqj_longer ? i : j;
       size_t s2 = is_seqj_longer ? j : i;
       double l1 = seq[s1].size(), l2 = seq[s2].size();
-      if (verbose) {
-        mutex.lock();
-        OPTIMOTU_COUT << "#### seq " << i << " (l1=" << l1 << ") and "
-                  << j << " (l2=" << l2 <<")####" << std::endl;
-        mutex.unlock();
-      }
+      OPTIMOTU_DEBUG(
+        2,
+        << "#### seq " << i << " (l1=" << l1 << ") and "
+        << j << " (l2=" << l2 <<")####" << std::endl
+      );
 
       double sim_threshold = 1.0 - threshold; // compiler can probably do this?
       if (l1/l2 < sim_threshold) continue;
@@ -80,26 +79,33 @@ void HybridSplitClusterWorker::operator()(std::size_t begin, std::size_t end) {
       }
       if (d < 1.0) ++my_aligned;
 
-      if (verbose) {
-        mutex.lock();
-        OPTIMOTU_COUT << "distance=" << d
-                      << std::endl;
-        mutex.unlock();
-      }
+      OPTIMOTU_DEBUG(
+        2,
+        << " distance=" << d
+        << std::endl
+      );
       if (d < threshold) (*my_algo)(j, i, d);
+      OPTIMOTU_DEBUG(
+        2,
+        << "thread" << begin
+        << ": finished " << j
+        << " and " << i
+        << "\n" << std::endl
+      );
       RcppThread::checkUserInterrupt();
     }
   }
   mutex.lock();
-  if (verbose) std::cout << "thread " << begin << " ready to merge" << std::endl;
+  OPTIMOTU_DEBUG(1, << "thread" << begin << " ready to merge" << std::endl);
   _aligned += my_aligned;
   _prealigned += my_prealigned;
   mutex.unlock();
   my_algo->merge_into_parent();
-  if (verbose) std::cout << "thread " << begin << " done" << std::endl;
+  OPTIMOTU_DEBUG(1, << "thread" << begin << " done" << std::endl);
 }
 
-void HybridConcurrentClusterWorker::operator()(std::size_t begin, std::size_t end) {
+template <int verbose>
+void HybridConcurrentClusterWorker<verbose>::operator()(std::size_t begin, std::size_t end) {
   double n = seq.size();
   double m = (n*n - 3.0*n + 2.0)/2.0;
   size_t my_prealigned = 0;
@@ -121,32 +127,37 @@ void HybridConcurrentClusterWorker::operator()(std::size_t begin, std::size_t en
     begin_i = round(1.5 + 0.5*sqrt(9.0 + 8.0*((m*begin)/threads - 1.0)));
   }
   size_t end_i   = round(1.5 + 0.5*sqrt(9.0 + 8.0*((m*end)/threads - 1.0)));
-  if (verbose) {
-    mutex.lock();
-    std::cout << "HybridConcurrent thread " << begin << " entered; sequences [" <<
-      begin_i << ", "<< end_i << ")" << std::endl;
-    mutex.unlock();
-  }
+  OPTIMOTU_DEBUG(
+    1,
+    << "HybridConcurrent thread " << begin
+    << " entered; sequences [" << begin_i
+    << ", "<< end_i << ")" << std::endl
+  );
   for (size_t i = begin_i; i < end_i; i++) {
     for (size_t j = 0; j < i; j++) {
-      // mutex.lock();
-      // std::cout << "Thread " << begin
-      //           << ": seqs " << j
-      //           << " and " << i
-      //           << std::endl;
-      // mutex.unlock();
+      OPTIMOTU_DEBUG(
+        2,
+        << "Thread " << begin
+        << ": seqs " << j
+        << " and " << i
+        << std::endl
+      );
       double threshold = clust_algo.max_relevant(i, j);
-      // mutex.lock();
-      // std::cout << "Thread " << begin
-      //           << ": max relevant=" << threshold
-      //                   << std::endl;
-      // mutex.unlock();
+      OPTIMOTU_DEBUG(
+        2,
+        << "Thread " << begin
+        << ": max relevant=" << threshold
+        << std::endl
+      );
       bool is_seqj_longer = seq[j].size() > seq[i].size();
       size_t s1 = is_seqj_longer ? i : j;
       size_t s2 = is_seqj_longer ? j : i;
       double l1 = seq[s1].size(), l2 = seq[s2].size();
-      // Rcpp::Rcout << "#### seq " << i << " (l1=" << l1 << ") and "
-      //             << j << " (l2=" << l2 <<")####" << std::endl;
+      OPTIMOTU_DEBUG(
+        2,
+        << "#### seq " << i << " (l1=" << l1 << ") and "
+        << j << " (l2=" << l2 <<")####" << std::endl
+      );
 
       double sim_threshold = 1.0 - threshold; // compiler can probably do this?
       if (l1/l2 < sim_threshold) continue;
@@ -160,75 +171,46 @@ void HybridConcurrentClusterWorker::operator()(std::size_t begin, std::size_t en
         int min_k = -(int)ceil((l1 - l2 * sim_threshold) / sim_threshold_plus_1);
         wfa_aligner.setHeuristicBandedStatic(min_k, max_k);
         wfa_aligner.setMaxAlignmentSteps((int)maxd1 + 1);
-        // std::cout << "wfa_aligner min_k=" << min_k
-        //           << " max_k=" << max_k
-        //           << " max score=" << (int)maxd1 + 1
-        //           << std::endl;
+        OPTIMOTU_DEBUG(
+          2,
+          << "wfa_aligner min_k=" << min_k
+          << " max_k=" << max_k
+          << " max score=" << (int)maxd1 + 1
+          << std::endl
+        );
         d = distance_wfa2(seq[s1], seq[s2], wfa_aligner);
       } else {
         ed_aligner.k = (int)maxd1 + 1;
         d = distance_edlib(seq[s1], seq[s2], ed_aligner);
       }
       if (d < 1.0) ++my_aligned;
-      // mutex.lock();
-      // std::cout << "Thread " << begin
-      //           << ": distance=" << d
-      //           << std::endl;
-      // mutex.unlock();
+      OPTIMOTU_DEBUG(
+        2,
+        << "Thread " << begin
+        << ": distance=" << d
+        << std::endl
+      );
       if (d < threshold) clust_algo(j, i, d);
-      // mutex.lock();
-      // std::cout << "Thread " << begin
-      //           << ": finished " << j
-      //           << " and " << i
-      //           << std::endl;
-      // mutex.unlock();
+      OPTIMOTU_DEBUG(
+        2,
+        << "Thread " << begin
+        << ": finished " << j
+        << " and " << i
+        << "\n" << std::endl
+      );
       RcppThread::checkUserInterrupt();
     }
   }
   mutex.lock();
   _aligned += my_aligned;
   _prealigned += my_prealigned;
-  // std::cout << "Exiting thread " << begin << std::endl;
+  OPTIMOTU_DEBUG(1, << "Exiting thread " << begin << std::endl);
   mutex.unlock();
 }
 
-// std::unique_ptr<SingleClusterAlgorithm> create_cluster_algorithm(
-//     const std::string &method,
-//     const DistanceConverter &dconv,
-//     init_matrix_t &im,
-//     bool do_binary_search,
-//     int fill_type
-// ) {
-//   if (method == "matrix") {
-//     if (do_binary_search) {
-//       switch (fill_type) {
-//       case LINEAR_FILL:
-//         return std::make_unique<ClusterMatrix<true, LINEAR_FILL, internal_matrix_t>>(dconv, im);
-//       case BINARY_FILL:
-//         return std::make_unique<ClusterMatrix<true, BINARY_FILL, internal_matrix_t>>(dconv, im);
-//       case TOPDOWN_FILL:
-//         return std::make_unique<ClusterMatrix<true, TOPDOWN_FILL, internal_matrix_t>>(dconv, im);
-//       default:
-//         Rcpp::stop("unknown fill type");
-//       }
-//     } else {
-//       switch (fill_type) {
-//       case LINEAR_FILL:
-//         return std::make_unique<ClusterMatrix<false, LINEAR_FILL, internal_matrix_t>>(dconv, im);
-//       case BINARY_FILL:
-//         return std::make_unique<ClusterMatrix<false, BINARY_FILL, internal_matrix_t>>(dconv, im);
-//       case TOPDOWN_FILL:
-//         return std::make_unique<ClusterMatrix<false, TOPDOWN_FILL, internal_matrix_t>>(dconv, im);
-//       default:
-//         Rcpp::stop("unknown fill type");
-//       }
-//     }
-//   } else if (method == "index") {
-//     return std::make_unique<ClusterIndexedMatrix<internal_matrix_t>>(dconv, im);
-//   } else if (method == "tree") {
-//     return std::make_unique<ClusterTree>(dconv, im);
-//   } else {
-//     Rcpp::stop("unknown cluster method");
-//   }
-// }
-
+template class HybridSplitClusterWorker<0>;
+template class HybridSplitClusterWorker<1>;
+template class HybridSplitClusterWorker<2>;
+template class HybridConcurrentClusterWorker<0>;
+template class HybridConcurrentClusterWorker<1>;
+template class HybridConcurrentClusterWorker<2>;