+++ src/RcppExports.cpp
@@ -291,7 +291,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // seq_cluster_single
-Rcpp::RObject seq_cluster_single(const Rcpp::CharacterVector& seq, const Rcpp::List dist_config, const Rcpp::List threshold_config, const Rcpp::List clust_config, const Rcpp::List parallel_config, const std::string output_type, const bool verbose);
+Rcpp::RObject seq_cluster_single(const Rcpp::CharacterVector& seq, const Rcpp::List dist_config, const Rcpp::List threshold_config, const Rcpp::List clust_config, const Rcpp::List parallel_config, const std::string output_type, const int verbose);
 RcppExport SEXP _optimotu_seq_cluster_single(SEXP seqSEXP, SEXP dist_configSEXP, SEXP threshold_configSEXP, SEXP clust_configSEXP, SEXP parallel_configSEXP, SEXP output_typeSEXP, SEXP verboseSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -302,13 +302,13 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const Rcpp::List >::type clust_config(clust_configSEXP);
     Rcpp::traits::input_parameter< const Rcpp::List >::type parallel_config(parallel_configSEXP);
     Rcpp::traits::input_parameter< const std::string >::type output_type(output_typeSEXP);
-    Rcpp::traits::input_parameter< const bool >::type verbose(verboseSEXP);
+    Rcpp::traits::input_parameter< const int >::type verbose(verboseSEXP);
     rcpp_result_gen = Rcpp::wrap(seq_cluster_single(seq, dist_config, threshold_config, clust_config, parallel_config, output_type, verbose));
     return rcpp_result_gen;
 END_RCPP
 }
 // seq_cluster_multi
-Rcpp::List seq_cluster_multi(const Rcpp::CharacterVector& seq, const Rcpp::ListOf<Rcpp::CharacterVector> which, const Rcpp::List dist_config, const Rcpp::List threshold_config, const Rcpp::List clust_config, const Rcpp::List parallel_config, const std::string output_type, const bool verbose);
+Rcpp::List seq_cluster_multi(const Rcpp::CharacterVector& seq, const Rcpp::ListOf<Rcpp::CharacterVector> which, const Rcpp::List dist_config, const Rcpp::List threshold_config, const Rcpp::List clust_config, const Rcpp::List parallel_config, const std::string output_type, const int verbose);
 RcppExport SEXP _optimotu_seq_cluster_multi(SEXP seqSEXP, SEXP whichSEXP, SEXP dist_configSEXP, SEXP threshold_configSEXP, SEXP clust_configSEXP, SEXP parallel_configSEXP, SEXP output_typeSEXP, SEXP verboseSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -320,7 +320,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const Rcpp::List >::type clust_config(clust_configSEXP);
     Rcpp::traits::input_parameter< const Rcpp::List >::type parallel_config(parallel_configSEXP);
     Rcpp::traits::input_parameter< const std::string >::type output_type(output_typeSEXP);
-    Rcpp::traits::input_parameter< const bool >::type verbose(verboseSEXP);
+    Rcpp::traits::input_parameter< const int >::type verbose(verboseSEXP);
     rcpp_result_gen = Rcpp::wrap(seq_cluster_multi(seq, which, dist_config, threshold_config, clust_config, parallel_config, output_type, verbose));
     return rcpp_result_gen;
 END_RCPP