+++ src/Wfa2ClusterWorker.cpp
@@ -8,13 +8,13 @@ Wfa2ClusterWorker::Wfa2ClusterWorker(
   const std::uint8_t threads,
   const int match, const int mismatch,
   const int gap_open, const int gap_extend,
-  const int gap_open2, const int gap_extend2,
-  bool verbose
-) : AlignClusterWorker(seq, clust_algo, threads, verbose),
+  const int gap_open2, const int gap_extend2
+) : AlignClusterWorker(seq, clust_algo, threads),
 match(match), mismatch(mismatch), gap_open(gap_open), gap_extend(gap_extend),
 gap_open2(gap_open2), gap_extend2(gap_extend2) {};
 
-void Wfa2SplitClusterWorker::operator()(std::size_t begin, std::size_t end) {
+template<int verbose>
+void Wfa2SplitClusterWorker<verbose>::operator()(std::size_t begin, std::size_t end) {
   double n = seq.size();
   double m = (n*n - 3.0*n + 2.0)/2.0;
   size_t my_prealigned = 0;
@@ -31,26 +31,33 @@ void Wfa2SplitClusterWorker::operator()(std::size_t begin, std::size_t end) {
     begin_i = round(1.5 + 0.5*sqrt(9.0 + 8.0*((m*begin)/threads - 1.0)));
   }
   size_t end_i   = round(1.5 + 0.5*sqrt(9.0 + 8.0*((m*end)/threads - 1.0)));
-  if (verbose) {
-    mutex.lock();
-    std::cout << "Wfa2Split thread " << begin << " entered; sequences [" <<
-      begin_i << ", "<< end_i << ")" << std::endl;
-    mutex.unlock();
-  }
+  OPTIMOTU_DEBUG(
+    1,
+    << "Wfa2Split thread " << begin
+    << " entered; sequences [" << begin_i
+    << ", "<< end_i << ")" << std::endl
+  );
   for (size_t i = begin_i; i < end_i; i++) {
     for (size_t j = 0; j < i; j++) {
       double threshold = my_algo->max_relevant(i, j);
-      // RcppThread::Rcout << "seqs " << j
-      //                   << " and " << i
-      //                   << " max relevant=" << threshold
-      //                   << std::endl;
+      OPTIMOTU_DEBUG(
+        2,
+        << "thread" << begin
+        << ": seqs " << j
+        << " and " << i
+        << " max relevant=" << threshold
+        << std::endl
+      );
 
       bool is_seqj_longer = seq[j].size() > seq[i].size();
       size_t s1 = is_seqj_longer ? i : j;
       size_t s2 = is_seqj_longer ? j : i;
       double l1 = seq[s1].size(), l2 = seq[s2].size();
-      // Rcpp::Rcout << "#### seq " << i << " (l1=" << l1 << ") and "
-      //             << j << " (l2=" << l2 <<")####" << std::endl;
+      OPTIMOTU_DEBUG(
+        2,
+        << "#### seq " << i << " (l1=" << l1 << ") and "
+        << j << " (l2=" << l2 <<")####" << std::endl
+      );
 
       double sim_threshold = 1.0 - threshold; // compiler can probably do this?
       if (l1/l2 < sim_threshold) continue;
@@ -64,22 +71,26 @@ void Wfa2SplitClusterWorker::operator()(std::size_t begin, std::size_t end) {
       double d = distance_wfa2(seq[s1], seq[s2], wfa_aligner);
       if (d < 1.0) ++my_aligned;
 
-      // std::cout << "distance=" << d
-      //                   << std::endl;
+      OPTIMOTU_DEBUG(
+        2,
+        << " distance=" << d
+        << std::endl
+      );
       if (d < threshold) (*my_algo)(j, i, d);
       RcppThread::checkUserInterrupt();
     }
   }
   mutex.lock();
-  if (verbose) std::cout << "thread " << begin << " ready to merge" << std::endl;
+  OPTIMOTU_DEBUG(1, << "thread " << begin << " ready to merge" << std::endl);
   _aligned += my_aligned;
   _prealigned += my_prealigned;
   mutex.unlock();
   my_algo->merge_into_parent();
-  if (verbose) std::cout << "thread " << begin << " done" << std::endl;
+  OPTIMOTU_DEBUG(1, << "thread " << begin << " done" << std::endl);
 }
 
-void Wfa2ConcurrentClusterWorker::operator()(std::size_t begin, std::size_t end) {
+template <int verbose>
+void Wfa2ConcurrentClusterWorker<verbose>::operator()(std::size_t begin, std::size_t end) {
   double n = seq.size();
   double m = (n*n - 3.0*n + 2.0)/2.0;
   size_t my_prealigned = 0;
@@ -95,32 +106,37 @@ void Wfa2ConcurrentClusterWorker::operator()(std::size_t begin, std::size_t end)
     begin_i = round(1.5 + 0.5*sqrt(9.0 + 8.0*((m*begin)/threads - 1.0)));
   }
   size_t end_i   = round(1.5 + 0.5*sqrt(9.0 + 8.0*((m*end)/threads - 1.0)));
-  if (verbose) {
-    mutex.lock();
-    std::cout << "HybridConcurrent thread " << begin << " entered; sequences [" <<
-      begin_i << ", "<< end_i << ")" << std::endl;
-    mutex.unlock();
-  }
+  OPTIMOTU_DEBUG(
+    1,
+    << "Wfa2Concurrent thread " << begin
+    << " entered; sequences [" << begin_i
+    << ", "<< end_i << ")" << std::endl
+  );
   for (size_t i = begin_i; i < end_i; i++) {
     for (size_t j = 0; j < i; j++) {
-      // mutex.lock();
-      // std::cout << "Thread " << begin
-      //           << ": seqs " << j
-      //           << " and " << i
-      //           << std::endl;
-      // mutex.unlock();
+      OPTIMOTU_DEBUG(
+        2,
+        << "Thread " << begin
+        << ": seqs " << j
+        << " and " << i
+        << std::endl
+      );
       double threshold = clust_algo.max_relevant(i, j);
-      // mutex.lock();
-      // std::cout << "Thread " << begin
-      //           << ": max relevant=" << threshold
-      //                   << std::endl;
-      // mutex.unlock();
+      OPTIMOTU_DEBUG(
+        2,
+        << "Thread " << begin
+        << ": max relevant=" << threshold
+        << std::endl
+      );
       bool is_seqj_longer = seq[j].size() > seq[i].size();
       size_t s1 = is_seqj_longer ? i : j;
       size_t s2 = is_seqj_longer ? j : i;
       double l1 = seq[s1].size(), l2 = seq[s2].size();
-      // Rcpp::Rcout << "#### seq " << i << " (l1=" << l1 << ") and "
-      //             << j << " (l2=" << l2 <<")####" << std::endl;
+      OPTIMOTU_DEBUG(
+        2,
+        << "#### seq " << i << " (l1=" << l1 << ") and "
+        << j << " (l2=" << l2 <<")####" << std::endl
+      );
 
       double sim_threshold = 1.0 - threshold; // compiler can probably do this?
       if (l1/l2 < sim_threshold) continue;
@@ -131,30 +147,42 @@ void Wfa2ConcurrentClusterWorker::operator()(std::size_t begin, std::size_t end)
       int min_k = -(int)ceil((l1 - l2 * sim_threshold) / sim_threshold_plus_1);
       wfa_aligner.setHeuristicBandedStatic(min_k, max_k);
       wfa_aligner.setMaxAlignmentSteps((int)maxd1 + 1);
-      // std::cout << "wfa_aligner min_k=" << min_k
-      //           << " max_k=" << max_k
-      //           << " max score=" << (int)maxd1 + 1
-      //           << std::endl;
+      OPTIMOTU_DEBUG(
+        2,
+        << "wfa_aligner min_k=" << min_k
+        << " max_k=" << max_k
+        << " max score=" << (int)maxd1 + 1
+        << std::endl
+      );
       double d = distance_wfa2(seq[s1], seq[s2], wfa_aligner);
       if (d < 1.0) ++my_aligned;
-      // mutex.lock();
-      // std::cout << "Thread " << begin
-      //           << ": distance=" << d
-      //           << std::endl;
-      // mutex.unlock();
+      OPTIMOTU_DEBUG(
+        2,
+        << "Thread " << begin
+        << " distance=" << d
+        << std::endl
+      );
       if (d < threshold) clust_algo(j, i, d);
-      // mutex.lock();
-      // std::cout << "Thread " << begin
-      //           << ": finished " << j
-      //           << " and " << i
-      //           << std::endl;
-      // mutex.unlock();
+      OPTIMOTU_DEBUG(
+        2,
+        << "Thread " << begin
+        << ": finished " << j
+        << " and " << i
+        << "\n" << std::endl
+      );
       RcppThread::checkUserInterrupt();
     }
   }
   mutex.lock();
   _aligned += my_aligned;
   _prealigned += my_prealigned;
-  // std::cout << "Exiting thread " << begin << std::endl;
+  OPTIMOTU_DEBUG(1, << "Exiting thread " << begin << std::endl);
   mutex.unlock();
 }
+
+template class Wfa2SplitClusterWorker<0>;
+template class Wfa2SplitClusterWorker<1>;
+template class Wfa2SplitClusterWorker<2>;
+template class Wfa2ConcurrentClusterWorker<0>;
+template class Wfa2ConcurrentClusterWorker<1>;
+template class Wfa2ConcurrentClusterWorker<2>;