+++ src/optimotu.h
@@ -2,9 +2,13 @@
 
 #ifdef OPTIMOTU_R
 
-#define OPTIMOTU_COUT Rcpp::Rcout
-#define OPTIMOTU_CERR Rcpp::Rcerr
+#include <RcppThread.h>
+#define OPTIMOTU_COUT RcppThread::Rcout
+#define OPTIMOTU_CERR RcppThread::Rcerr
 #define OPTIMOTU_STOP Rcpp::stop
+#define OPTIMOTU_DEBUG(level, s) do {\
+  if constexpr (level <= verbose) RcppThread::Rcerr s;\
+} while (false)
 
 #else
 
@@ -12,6 +16,9 @@
 #define OPTIMOTU_COUT std::cout
 #define OPTIMOTU_CERR std::cerr
 #define OPTIMOTU_STOP(s) do { std::cerr << (s); std::exit(1); } while (false)
+#define OPTIMOTU_DEBUG(level, s) do {\
+  if constexpr (level <= verbose) std::cerr s;\
+} while (false)
 
 #endif // OPTIMOTU_R